+++ src/main/java/teamkim/barrier_free_kiosk_spring/service/KioskService.java
@@ -3,7 +3,7 @@ package teamkim.barrier_free_kiosk_spring.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import teamkim.barrier_free_kiosk_spring.dto.MoveInReportReqDto;
+import teamkim.barrier_free_kiosk_spring.dto.request.MoveInReportReqDto;
 import teamkim.barrier_free_kiosk_spring.entity.Address;
 import teamkim.barrier_free_kiosk_spring.entity.MoveInReportLog;
 import teamkim.barrier_free_kiosk_spring.entity.User;
@@ -11,7 +11,7 @@ import teamkim.barrier_free_kiosk_spring.exception.CustomException;
 import teamkim.barrier_free_kiosk_spring.exception.ExceptionCode;
 import teamkim.barrier_free_kiosk_spring.repository.LogRepository;
 import teamkim.barrier_free_kiosk_spring.repository.UserRepository;
-import teamkim.barrier_free_kiosk_spring.dto.ResidentRegistrationReqDto;
+import teamkim.barrier_free_kiosk_spring.dto.request.ResidentRegistrationReqDto;
 import teamkim.barrier_free_kiosk_spring.entity.ResidentRegistrationLog;
 
 @Service
